feature: Criação de lógica para não permitir que um usuário edite ou exclua o projeto de outro usuário
- Foi adicionada uma verificação nos Services de edição e exclusão de projetos para que não permita alterar o projeto caso seja de outro usuário
- Foi criada uma exception para esses casos onde o usuário tenta alterar um projeto de outro usuário
- Foi criado um Dto específico para a rota de exclusão do projeto

+++ src/Projects/Domain/Service/DeleteProject.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace OrangePortfolio\Projects\Domain\Service;
 
+use OrangePortfolio\Projects\Domain\Dto\DeleteProjectDto;
+use OrangePortfolio\Projects\Domain\Exception\AnotherUsersProject;
 use OrangePortfolio\Projects\Domain\Repository\ProjectRepositoryInterface;
 
 class DeleteProject
@@ -12,9 +14,13 @@ class DeleteProject
     {
     }
 
-    public function delete(int $id): void
+    public function delete(DeleteProjectDto $deleteProjectDto): void
     {
-        $project = $this->projectRepository->getById($id);
+        $project = $this->projectRepository->getById($deleteProjectDto->projectId);
+        if ($project->getUser()->getId() !== $deleteProjectDto->userId) {
+            throw AnotherUsersProject::fromId($deleteProjectDto->userId);
+        }
+
         $this->projectRepository->delete($project);
     }
 }
